requirements.txt
pandas
advertools
google-generativeai
gradio
python-dotenv
markdownify
mdformat